Routing feature was added, navigation menu is separated to be reusable component, 'a' tags are replaced by Link tag. Also page of orders was initially created
