+++ src/App.js
@@ -2,7 +2,9 @@ import React,{useState} from "react";
 import 'normalize.css';
 import './styles/index.scss';
 import ModelAuthorization from "./component/ModelAuthorization";
-import StoragePage from "./component/StoragePage";
+import StoragePage from "./pages/StoragePage";
+import {BrowserRouter, Route, Routes} from "react-router-dom";
+import OrdersPage from "./pages/OrdersPage";
 
 function App() {
   const admin = {login:'admin', password:'admin'};
@@ -11,11 +13,18 @@ function App() {
     setAuthorize(!authorizat);
   }
   return (
-    <div className="App">
-      {authorizat ? 
-        <StoragePage isAuthorizetion={isAuthorizetion}/>  : 
-        <ModelAuthorization isAuthorizetion={isAuthorizetion} admin={admin}/>}
-    </div>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={(<div className="App">
+            {authorizat ?
+                <StoragePage isAuthorizetion={isAuthorizetion}/>  :
+                <ModelAuthorization isAuthorizetion={isAuthorizetion} admin={admin}/>}
+          </div>)}/>
+          <Route path="/orders" element={<OrdersPage/>}/>
+          <Route path="/storage" element={<StoragePage/>}/>
+        </Routes>
+      </BrowserRouter>
+
   );
 }
