--- src/component/StoragePage.jsx
@@ -1,112 +1,0 @@
-import React,{useState} from 'react';
-import AddModalVIew from './AddModalVIew';
-import AddSupleModalView from './AddSupleModalView';
-import IngredientTable from './IngredientTable';
-
-const StoragePage = ({isAuthorizetion}) => {
-	const [modalActive,setModalActive] = useState(false);
-	const [supleModelActive,setSupleModalActive] = useState(false);
-  const exit =(e) => {
-    e.preventDefault();
-    isAuthorizetion();
-  }
-  const tf =(e) => {
-		e.preventDefault();
-	}
-  return (
-    <div className='outer-wrapper main'>
-      <nav className="nav-bar">
-        <ul className="nav-bar_list">
-          <li className="item-list selected">
-						<svg className='item-list_icon' width="56" height="55" viewBox="0 0 56 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-						<path d="M48.4279 3.73083C48.2592 3.30565 47.9682 2.94129 47.5925 2.68476C47.2168 2.42824 46.7736 2.29131 46.32 2.29166H10.0164C9.56289 2.29131 9.11967 2.42824 8.74395 2.68476C8.36822 2.94129 8.07723 3.30565 7.90852 3.73083L3.37057 15.1892C3.26424 15.4606 3.20958 15.7498 3.20947 16.0417V50.4167C3.20947 51.0245 3.44852 51.6073 3.87404 52.0371C4.29956 52.4669 4.87668 52.7083 5.47845 52.7083H50.858C51.4598 52.7083 52.0369 52.4669 52.4624 52.0371C52.8879 51.6073 53.127 51.0245 53.127 50.4167V16.0417C53.1269 15.7498 53.0722 15.4606 52.9659 15.1892L48.4279 3.73083ZM11.5525 6.875H44.7839L47.5067 13.75H8.82973L11.5525 6.875ZM7.74743 48.125V18.3333H48.589V48.125H7.74743ZM25.8992 41.25C25.8992 41.8578 25.6602 42.4407 25.2347 42.8705C24.8092 43.3002 24.232 43.5417 23.6303 43.5417H14.5544C13.9526 43.5417 13.3755 43.3002 12.9499 42.8705C12.5244 42.4407 12.2854 41.8578 12.2854 41.25C12.2854 40.6422 12.5244 40.0593 12.9499 39.6295C13.3755 39.1998 13.9526 38.9583 14.5544 38.9583H23.6303C24.232 38.9583 24.8092 39.1998 25.2347 39.6295C25.6602 40.0593 25.8992 40.6422 25.8992 41.25Z" fill="black"/>
-						</svg>
-						<span className='item-list_text'>
-							<a href='/storage/' 
-							onClick={tf}
-							className='item-list_link'>Storage</a>
-						</span>
-					</li>
-          <li className="item-list">
-						<svg className='item-list_icon' width="53" height="53" viewBox="0 0 53 53" fill="none" xmlns="http://www.w3.org/2000/svg">
-						<path d="M0 1.76667C0 0.790964 0.790963 0 1.76667 0H7.06666C7.87733 0 8.58397 0.551725 8.78058 1.33819L10.2127 7.06667H51.2333C51.7589 7.06667 52.2571 7.30066 52.5928 7.70508C52.9284 8.1095 53.0666 8.64235 52.9697 9.15891L47.6697 37.4256C47.5131 38.2612 46.7835 38.8667 45.9333 38.8667H42.4H17.6667H14.1333C13.2832 38.8667 12.5536 38.2612 12.3969 37.4256L7.10663 9.2107L5.68729 3.53333H1.76667C0.790963 3.53333 0 2.74237 0 1.76667ZM10.962 10.6L15.5995 35.3333H44.4671L49.1046 10.6H10.962ZM17.6667 38.8667C13.7639 38.8667 10.6 42.0305 10.6 45.9333C10.6 49.8361 13.7639 53 17.6667 53C21.5695 53 24.7333 49.8361 24.7333 45.9333C24.7333 42.0305 21.5695 38.8667 17.6667 38.8667ZM42.4 38.8667C38.4972 38.8667 35.3333 42.0305 35.3333 45.9333C35.3333 49.8361 38.4972 53 42.4 53C46.3028 53 49.4667 49.8361 49.4667 45.9333C49.4667 42.0305 46.3028 38.8667 42.4 38.8667ZM17.6667 42.4C19.6181 42.4 21.2 43.9819 21.2 45.9333C21.2 47.8847 19.6181 49.4667 17.6667 49.4667C15.7153 49.4667 14.1333 47.8847 14.1333 45.9333C14.1333 43.9819 15.7153 42.4 17.6667 42.4ZM42.4 42.4C44.3514 42.4 45.9333 43.9819 45.9333 45.9333C45.9333 47.8847 44.3514 49.4667 42.4 49.4667C40.4486 49.4667 38.8667 47.8847 38.8667 45.9333C38.8667 43.9819 40.4486 42.4 42.4 42.4Z" fill="black"/>
-						</svg>
-						<span className='item-list_text'>
-							<a href='/orders/'
-							onClick={tf}
-							className='item-list_link'>Orders</a>
-						</span>
-					</li>
-          <li className="item-list">
-						<svg className='item-list_icon' width="57" height="57" viewBox="0 0 57 57" fill="none" xmlns="http://www.w3.org/2000/svg">
-						<path d="M24.9375 49.875C24.9375 49.875 21.375 49.875 21.375 46.3125C21.375 42.75 24.9375 32.0625 39.1875 32.0625C53.4375 32.0625 57 42.75 57 46.3125C57 49.875 53.4375 49.875 53.4375 49.875H24.9375Z" fill="black"/>
-						<path d="M39.1875 28.5C45.09 28.5 49.875 23.715 49.875 17.8125C49.875 11.91 45.09 7.125 39.1875 7.125C33.285 7.125 28.5 11.91 28.5 17.8125C28.5 23.715 33.285 28.5 39.1875 28.5Z" fill="black"/>
-						<path  d="M18.5833 49.875C18.0817 48.8629 17.8125 47.6632 17.8125 46.3125C17.8125 41.4837 20.2316 36.5176 24.7093 33.0599C22.7505 32.4333 20.4669 32.0625 17.8125 32.0625C3.5625 32.0625 0 42.75 0 46.3125C0 49.875 3.5625 49.875 3.5625 49.875H18.5833Z" fill="black"/>
-						<path d="M16.0312 28.5C20.95 28.5 24.9375 24.5125 24.9375 19.5938C24.9375 14.675 20.95 10.6875 16.0312 10.6875C11.1125 10.6875 7.125 14.675 7.125 19.5938C7.125 24.5125 11.1125 28.5 16.0312 28.5Z" fill="black"/>
-						</svg>
-						<span className='item-list_text'>
-							<a href='/profiles/'
-							onClick={tf}
-							className='item-list_link'>Profiles</a>
-						</span>
-					</li>
-          <li className="item-list">
-						<svg className='item-list_icon' width="55" height="55" viewBox="0 0 55 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-						<path d="M37.8125 6.875C37.8125 4.97652 39.3515 3.4375 41.25 3.4375H48.125C50.0235 3.4375 51.5625 4.97652 51.5625 6.875V48.125H53.2812C54.2305 48.125 55 48.8945 55 49.8438C55 50.793 54.2305 51.5625 53.2812 51.5625H1.71875C0.769511 51.5625 0 50.793 0 49.8438C0 48.8945 0.769511 48.125 1.71875 48.125H3.4375V37.8125C3.4375 35.914 4.97652 34.375 6.875 34.375H13.75C15.6485 34.375 17.1875 35.914 17.1875 37.8125V48.125H20.625V24.0625C20.625 22.164 22.164 20.625 24.0625 20.625H30.9375C32.836 20.625 34.375 22.164 34.375 24.0625V48.125H37.8125V6.875ZM41.25 48.125H48.125V6.875H41.25V48.125ZM30.9375 48.125V24.0625H24.0625V48.125H30.9375ZM13.75 48.125V37.8125H6.875V48.125H13.75Z" fill="black"/>
-						</svg>
-						<span className='item-list_text'>
-						<a href='/stats/'
-							onClick={tf}
-							className='item-list_link'>Stats</a>
-						</span>
-					</li>
-					<li className="item-list" onClick={exit}>
-						<svg width="40" height="40" viewBox="0 0 16 17" fill="none" xmlns="http://www.w3.org/2000/svg">
-						<path d="M0.505588 1.23397C0.941099 0.810897 1.6472 0.810897 2.08271 1.23397L7.98533 6.96793L13.8879 1.23397C14.3234 0.810897 15.0296 0.810897 15.4651 1.23397C15.9006 1.65703 15.9006 2.34296 15.4651 2.76603L9.56245 8.5L15.4651 14.234C15.9006 14.657 15.9006 15.343 15.4651 15.766C15.0296 16.1891 14.3234 16.1891 13.8879 15.766L7.98533 10.0321L2.08271 15.766C1.6472 16.1891 0.941099 16.1891 0.505588 15.766C0.0700773 15.343 0.0700773 14.657 0.505588 14.234L6.4082 8.5L0.505588 2.76603C0.0700773 2.34296 0.0700773 1.65703 0.505588 1.23397Z" fill="black"/>
-						</svg>
-						<span className='item-list_text'>
-						<a href='/exit/'
-							className='item-list_link'>Exit</a>
-						</span>
-					</li>
-        </ul>
-      </nav>
-			<AddModalVIew active={modalActive} setActive={setModalActive}/> 
-			<AddSupleModalView activeSuple={supleModelActive} setSupleModalActive={setSupleModalActive}/>
-			<div className='storage'>
-			<h2 className="storage-title">Storage</h2>
-			<div className="storage-mod">
-				<button
-					className='storage-btn add_btn'
-					onClick={() => setModalActive(true)}
-				>
-					<svg className='storage-mod_icon' width="19" height="15" viewBox="0 0 19 15" fill="none" xmlns="http://www.w3.org/2000/svg">
-					<path d="M9.12152 0.0714264C9.74046 0.0714264 10.2422 0.487162 10.2422 0.999998V6.57143H16.9663C17.5853 6.57143 18.087 6.98716 18.087 7.5C18.087 8.01283 17.5853 8.42857 16.9663 8.42857H10.2422V14C10.2422 14.5128 9.74046 14.9286 9.12152 14.9286C8.50258 14.9286 8.00083 14.5128 8.00083 14V8.42857H1.2767C0.657756 8.42857 0.156006 8.01283 0.156006 7.5C0.156006 6.98716 0.657756 6.57143 1.2767 6.57143H8.00083V0.999998C8.00083 0.487162 8.50258 0.0714264 9.12152 0.0714264Z" fill="black"/>
-					</svg>
-					<span className='storage-btn_title'>Add</span>
-				</button>
-				<button
-					className='storage-btn filter_btn'>
-					<svg className='storage-mod_icon' width="35" height="31" viewBox="0 0 35 31" fill="none" xmlns="http://www.w3.org/2000/svg">
-					<path d="M12.9961 20.3438C12.9961 19.8087 13.4705 19.375 14.0557 19.375H20.4131C20.9983 19.375 21.4727 19.8087 21.4727 20.3438C21.4727 20.8788 20.9983 21.3125 20.4131 21.3125H14.0557C13.4705 21.3125 12.9961 20.8788 12.9961 20.3438Z" fill="black"/>
-					<path d="M8.75781 14.5312C8.75781 13.9962 9.2322 13.5625 9.81738 13.5625H24.6514C25.2366 13.5625 25.7109 13.9962 25.7109 14.5312C25.7109 15.0663 25.2366 15.5 24.6514 15.5H9.81738C9.2322 15.5 8.75781 15.0663 8.75781 14.5312Z" fill="black"/>
-					<path d="M4.51953 8.71875C4.51953 8.18372 4.99392 7.75 5.5791 7.75H28.8896C29.4748 7.75 29.9492 8.18372 29.9492 8.71875C29.9492 9.25378 29.4748 9.6875 28.8896 9.6875H5.5791C4.99392 9.6875 4.51953 9.25378 4.51953 8.71875Z" fill="black"/>
-					</svg>
-					<span className='storage-btn_title'>Filter</span>
-				</button>
-				<button
-					className='storage-btn delete_btn'>
-					<svg className='storage-mod_icon' width="16" height="17" viewBox="0 0 16 17" fill="none" xmlns="http://www.w3.org/2000/svg">
-					<path d="M0.505588 1.23397C0.941099 0.810897 1.6472 0.810897 2.08271 1.23397L7.98533 6.96793L13.8879 1.23397C14.3234 0.810897 15.0296 0.810897 15.4651 1.23397C15.9006 1.65703 15.9006 2.34296 15.4651 2.76603L9.56245 8.5L15.4651 14.234C15.9006 14.657 15.9006 15.343 15.4651 15.766C15.0296 16.1891 14.3234 16.1891 13.8879 15.766L7.98533 10.0321L2.08271 15.766C1.6472 16.1891 0.941099 16.1891 0.505588 15.766C0.0700773 15.343 0.0700773 14.657 0.505588 14.234L6.4082 8.5L0.505588 2.76603C0.0700773 2.34296 0.0700773 1.65703 0.505588 1.23397Z" fill="black"/>
-					</svg>
-					<span className='storage-btn_title'>Delete</span>
-				</button>
-			</div>
-			<IngredientTable setSupleModalActive={setSupleModalActive}/>
-			</div>
-    </div>
-		
-    )
-}
-
-export default StoragePage
